Fixed styling and refactoring issues
