+++ src/scripts/data.js
@@ -21,18 +21,16 @@ const API = {
       headers: {
         "Content-Type": "application/json"
       }
-    }).then(API.getAndUpdate);
+    })
   },
   editJournalEntry: function(entry) {
-    fetch(`http://localhost:8088/collectionJournalEntries/${entry.id}`, {
+    return fetch(`http://localhost:8088/collectionJournalEntries/${entry.id}`, {
       method: "PUT",
       headers: {
         "Content-Type": "application/json"
       },
       body: JSON.stringify(entry)
-    })
-      .then(response => response.json())
-      .then(API.getAndUpdate);
+    });
   },
   getAndUpdate: function() {
     return API.getJournalEntries().then(entries_obj => {
